Add error handler for 410 errors and improve logging in various modules

+++ mielenosoitukset_fi/scripts/CL.py
@@ -1,6 +1,7 @@
 import os
 import time
 from datetime import datetime, timedelta
+from mielenosoitukset_fi.utils.logger import logger
 
 
 def remove_files_in_directory(directory):
@@ -12,14 +13,22 @@ def remove_files_in_directory(directory):
     directory : str
         The path to the directory where files will be removed.
     """
+    if not os.path.exists(directory):
+        logger.error(f"Directory does not exist: {directory}")
+        return
+
+    if not os.path.isdir(directory):
+        logger.error(f"Path is not a directory: {directory}")
+        return
+
     for filename in os.listdir(directory):
         file_path = os.path.join(directory, filename)
         try:
             if os.path.isfile(file_path):
                 os.unlink(file_path)
-                print(f"Removed file: {file_path}")
+                logger.info(f"Removed file: {file_path}")
         except Exception as e:
-            print(f"Failed to remove {file_path}. Reason: {e}")
+            logger.error(f"Failed to remove {file_path}. Reason: {e}")
 
 
 def main():
